# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuickLink API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/analytics/{shortCode}": {
      "get": {
        "operationId": "getAnalytics",
        "parameters": [
          { "name": "shortCode", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "range", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "from", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "to", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "country", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "device", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "referrer", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "ruleId", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Link analytics", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AnalyticsResponse" } } } } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyticsResponse": {
        "type": "object",
        "properties": {
          "url": { "type": "object" }, "window": { "type": "object" },
          "filters": { "type": "object" }, "analytics": { "type": "object" }
        },
        "required": ["url","window","analytics"]
      },
...
    }
  }
}